修改秒杀活动,新增超时取消订单逻辑

+++ backend/src/main/java/com/studynote/notes/model/dto/course/CreateCourseBody.java
@@ -3,6 +3,8 @@ package com.studynote.notes.model.dto.course;
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
 
+import java.math.BigDecimal;
+
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -13,7 +15,7 @@ public class CreateCourseBody {
     @NotBlank(message = "课程标题不能为空") String title;
     String description;                          // 可选
     String coverUrl;                             // 可选
-    @NotNull(message = "价格不能为空") @Min(value = 0, message = "价格不能为负") Long price;
+    @NotNull(message = "价格不能为空") @Min(value = 0, message = "价格不能为负") BigDecimal price;
     Integer status;                              // 可选，不传默认当 1 上架处理（自己定）
 
 }
