+++ backend/src/main/java/com/studynote/notes/model/dto/course/UpdateCourseBody.java
@@ -3,6 +3,8 @@ package com.studynote.notes.model.dto.course;
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
 
+import java.math.BigDecimal;
+
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -12,7 +14,7 @@ public class UpdateCourseBody {
     String title;
     String description;
     String coverUrl;
-    @Min(value = 0, message = "价格不能为负") Long price;   // 只加 @Min，不加 @NotNull（可选）
+    @Min(value = 0, message = "价格不能为负") BigDecimal price;   // 只加 @Min，不加 @NotNull（可选）
     Integer status;
 
 }
