+++ backend/src/main/java/com/studynote/notes/model/dto/seckillActivity/UpdateSeckillActivityBody.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
 import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -12,7 +13,7 @@ public class UpdateSeckillActivityBody {
     private Integer courseId;
 
     @Min(value = 0, message = "秒杀价不能为负")
-    private Long seckillPrice;
+    private BigDecimal seckillPrice;
 
     @Min(value = 1, message = "库存必须大于0")
     private Integer stock;
