+++ backend/src/main/java/com/studynote/notes/model/vo/seckill/SeckillActivityVO.java
@@ -2,6 +2,7 @@ package com.studynote.notes.model.vo.seckill;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -9,7 +10,7 @@ public class SeckillActivityVO {
 
     private Integer activityId;
     private Integer courseId;
-    private Long seckillPrice;
+    private BigDecimal seckillPrice;
     private Integer stock;
     private Date startTime;
     private Date endTime;
